requirements.txt
openai
python-dotenv
requests
supabase
airtable-python-wrapper